clean requirements
I finally managed to figure out environments in Python, so I cloned the repo and install only required libraries. this file is generated by pip freeze without the --all option, because in that case its started to reinstall pip, wheel and other installation libraries, making the process to hang up.